moved oanda connection file to shared project, updated to directly call the api rather then using library
library does not support getting symbols other then forex
