+++ Tradeinator.DataIngestion.Shared/Tradeinator.DataIngestion.Shared.csproj
@@ -8,6 +8,7 @@
 
     <ItemGroup>
       <PackageReference Include="Alpaca.Markets" Version="7.0.0" />
+      <PackageReference Include="GeriRemenyi.Oanda.V20.Sdk" Version="0.0.6" />
       <PackageReference Include="Serilog" Version="3.1.1" />
     </ItemGroup>
